+++ src/cli.ts
@@ -4,8 +4,19 @@ import { updateAll } from "./update-packages"
 
 const [,, ...args] = process.argv
 
-if(args.length > 0) {
-    console.log("Invalid arguments - we don't take any arguments (yet)")
+export function handleArgs(args: string[]) {
+    if(args.length > 0) {
+      console.log("Invalid arguments - we don't take any arguments (yet)")
+      return false;
+    }
+    return true;
+  }
+
+export function init(args: string[]){
+    handleArgs(args)
+    updateAll().then(() => console.log("update complete")).catch((error) => console.error(error))
 }
 
-updateAll().then(() => console.log("update complete")).catch((error) => console.error(error))
+init(args)
+
+